rebase and change scope interface
