+++ common/metrics/scope.go
@@ -22,22 +22,6 @@ package metrics
 
 import "github.com/uber-go/tally"
 
-const domainTag = "domain"
-
-// Tag is an interface to define metrics tags
-type Tag interface {
-	Key() string
-	Value() string
-}
-
-// Scope is an interface for metrics
-type Scope interface {
-	Counter(id int) tally.Counter
-	Gauge(id int) tally.Gauge
-	Timer(id int) tally.Timer
-	Tagged(tags ...Tag) Scope
-}
-
 type metricsScope struct {
 	scope tally.Scope
 	defs  map[int]metricDefinition
@@ -68,24 +52,4 @@ func (m *metricsScope) Tagged(tags ...Tag) Scope {
 		tagMap[tag.Key()] = tag.Value()
 	}
 	return newMetricsScope(m.scope.Tagged(tagMap), m.defs)
-}
-
-// DomainTag wraps a domain tag
-type DomainTag struct {
-	value string
-}
-
-// NewDomainTag returns a new domain tag
-func NewDomainTag(value string) DomainTag {
-	return DomainTag{value}
-}
-
-// Key returns the key of the domain tag
-func (d DomainTag) Key() string {
-	return domainTag
-}
-
-// Value returns the value of a domain tag
-func (d DomainTag) Value() string {
-	return d.value
 }
